+++ packages/flutter/lib/material.dart
@@ -53,7 +53,6 @@ export 'src/material/ink_highlight.dart';
 export 'src/material/ink_splash.dart';
 export 'src/material/ink_well.dart';
 export 'src/material/input.dart';
-export 'src/material/list.dart';
 export 'src/material/list_item.dart';
 export 'src/material/material.dart';
 export 'src/material/mergeable_material.dart';
