+++ packages/flutter/lib/src/material/two_level_list.dart
@@ -10,7 +10,6 @@ import 'icon.dart';
 import 'icons.dart';
 import 'icon_theme.dart';
 import 'icon_theme_data.dart';
-import 'list.dart';
 import 'list_item.dart';
 import 'theme.dart';
 import 'theme_data.dart';
