+++ packages/flutter/lib/src/rendering/viewport.dart
@@ -4,7 +4,6 @@
 
 import 'dart:ui' as ui show window;
 
-import 'package:flutter/foundation.dart';
 import 'package:vector_math/vector_math_64.dart';
 
 import 'box.dart';
@@ -335,134 +334,5 @@ class RenderViewport extends RenderViewportBase with RenderObjectWithChildMixin<
       return child.hitTest(result, position: transformed);
     }
     return false;
-  }
-}
-
-/// A render object that shows a subset of its children.
-///
-/// The children of a viewport can layout to a larger size along the viewport's
-/// [mainAxis] than the viewport itself. If that happens, only a subset of the
-/// children will be visible through the viewport. The subset of children that
-/// are visible can be controlled with the [paintOffset].
-///
-/// See also:
-///
-///  * [RenderList] (which arranges its children linearly)
-///  * [RenderGrid] (which arranges its children into tiles)
-///  * [RenderViewport] (which is easier to use with a single child)
-abstract class RenderVirtualViewport<T extends ContainerBoxParentDataMixin<RenderBox>>
-    extends RenderViewportBase with ContainerRenderObjectMixin<RenderBox, T>,
-                                    RenderBoxContainerDefaultsMixin<RenderBox, T> {
-  /// Initializes fields for subclasses.
-  ///
-  /// The [paintOffset] and [mainAxis] arguments must not be null.
-  RenderVirtualViewport({
-    int virtualChildCount,
-    LayoutCallback<BoxConstraints> callback,
-    Offset paintOffset: Offset.zero,
-    Axis mainAxis: Axis.vertical,
-    ViewportAnchor anchor: ViewportAnchor.start
-  }) : _virtualChildCount = virtualChildCount,
-       _callback = callback,
-       super(paintOffset, mainAxis, anchor);
-
-  /// The overall number of children this viewport could potentially display.
-  ///
-  /// If null, the viewport might display an unbounded number of children.
-  int get virtualChildCount => _virtualChildCount;
-  int _virtualChildCount;
-  set virtualChildCount(int value) {
-    if (_virtualChildCount == value)
-      return;
-    _virtualChildCount = value;
-    markNeedsLayout();
-  }
-
-  /// Called during [layout] to determine the render object's children.
-  ///
-  /// Typically the callback will mutate the child list appropriately, for
-  /// example so the child list contains only visible children.
-  LayoutCallback<BoxConstraints> get callback => _callback;
-  LayoutCallback<BoxConstraints> _callback;
-  set callback(LayoutCallback<BoxConstraints> value) {
-    if (value == _callback)
-      return;
-    _callback = value;
-    markNeedsLayout();
-  }
-
-  /// Throws an exception if asserts are enabled, unless the
-  /// [RenderObject.debugCheckingIntrinsics] flag is set.
-  ///
-  /// This is a convenience function for subclasses to call from their
-  /// intrinsic-sizing functions if they don't have a good way to generate the
-  /// numbers.
-  @protected
-  bool debugThrowIfNotCheckingIntrinsics() {
-    assert(() {
-      if (!RenderObject.debugCheckingIntrinsics) {
-        throw new FlutterError(
-          '$runtimeType does not support returning intrinsic dimensions.\n'
-          'Calculating the intrinsic dimensions would require walking the entire '
-          'child list, which cannot reliably and efficiently be done for render '
-          'objects that potentially generate their child list during layout.'
-        );
-      }
-      return true;
-    });
-    return true;
-  }
-
-  @override
-  double computeMinIntrinsicWidth(double height) {
-    assert(debugThrowIfNotCheckingIntrinsics());
-    return 0.0;
-  }
-
-  @override
-  double computeMaxIntrinsicWidth(double height) {
-    assert(debugThrowIfNotCheckingIntrinsics());
-    return 0.0;
-  }
-
-  @override
-  double computeMinIntrinsicHeight(double width) {
-    assert(debugThrowIfNotCheckingIntrinsics());
-    return 0.0;
-  }
-
-  @override
-  double computeMaxIntrinsicHeight(double width) {
-    assert(debugThrowIfNotCheckingIntrinsics());
-    return 0.0;
-  }
-
-  @override
-  bool hitTestChildren(HitTestResult result, { Point position }) {
-    return defaultHitTestChildren(result, position: position + -_effectivePaintOffset);
-  }
-
-  void _paintContents(PaintingContext context, Offset offset) {
-    defaultPaint(context, offset + _effectivePaintOffset);
-  }
-
-  @override
-  void paint(PaintingContext context, Offset offset) {
-    context.pushClipRect(needsCompositing, offset, Point.origin & size, _paintContents);
-  }
-
-  @override
-  Rect describeApproximatePaintClip(RenderObject child) => Point.origin & size;
-
-  // Workaround for https://github.com/dart-lang/sdk/issues/25232
-  @override
-  void applyPaintTransform(RenderBox child, Matrix4 transform) {
-    super.applyPaintTransform(child, transform);
-  }
-
-  @override
-  void debugFillDescription(List<String> description) {
-    super.debugFillDescription(description);
-    description.add('virtual child count: $virtualChildCount');
   }
 }
